move test_view_change_no_quorum to own file
